% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkEva.R
\name{BenchmarkEva}
\alias{BenchmarkEva}
\title{Title}
\usage{
BenchmarkEva(
  dataset = MLtestData,
  target = "judge",
  taskid = "MLtest",
  learnerlist = list(lrn("classif.rpart"), lrn("classif.featureless"))
)
}
\arguments{
\item{dataset}{A dataframe of Machine Learing test dataset}

\item{target}{A character of column name in dataset}

\item{taskid}{A character of taskid}

\item{learnerlist}{A vector of learners which would like to be compared}
}
\value{

}
\description{
Title
}
