% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCorrelation.R
\name{CorBetween}
\alias{CorBetween}
\title{Title}
\usage{
CorBetween(FunMatrix, outputname, width = 3000, height = 3000)
}
\arguments{
\item{outputname}{a character of output name}

\item{width}{a numeric of width}

\item{height}{a numeric of height}

\item{FeaMatrix}{a dataframe of feature matrix}
}
\value{
a dataframe with all correlation between each other which is bigger than 0.8;
and a hist plot
}
\description{
Title
}
