% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCorrelation.R
\name{CorPlot}
\alias{CorPlot}
\title{Title}
\usage{
CorPlot(
  FeaMatrix,
  outputname,
  width = 3000,
  height = 3000,
  corvismethod = "circle",
  order = "original"
)
}
\arguments{
\item{FeaMatrix}{a dataframe of feature matrix}

\item{outputname}{a character of output name}

\item{width}{a numeric of width}

\item{height}{a numeric of height}

\item{corvismethod}{a character of visulization method}

\item{order}{order to visualization,detail see corrplot}
}
\value{
a figure with png format
}
\description{
Title
}
