% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepKmerFeature.R
\name{Feature_Calcualtion}
\alias{Feature_Calcualtion}
\alias{PepFakeCode}
\title{Fragment feature calculation}
\usage{
PepFakeCode(peptideSet, coreN = parallel::detectCores(logical = F))
}
\arguments{
\item{peptideSet}{A vector a group of peptides}

\item{coreN}{The number of cores, default is all core in your machine}

\item{peptides}{A vector a group of peptides}

\item{JudgeMatrix}{A data.table with immune and non-immne info.}

\item{prop}{A proportion of fragment in all peptides/}

\item{fragLenSet}{A set of sliding window}

\item{featureSet}{A vector to select feature,default is NULL}

\item{Fake}{a logical TRUE or FALSE}

\item{FakeMatrix}{A dataframe of fakacode info, if Fake is ture,then this parameter need to be gaven. Colnames}

\item{CutStart}{a numeric of cut peptides from, default is NULL}

\item{Cutend}{a number of cut peptides to, default is NULL}
}
\description{
\code{pepFakecode} Rscript Coding peptide with specific rule of code.
\code{pepFrag} Rscript to generate Fragment of peptides.
\Code{pepFragSign} filtering all significant fragment
\code{PepFeatureDesc} pepFeature description
}
\examples{
peptideSetFake <- PepFakeCode(peptideSet)
}
