% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PepFeatureSelect.R
\name{FeatureSebyAuto}
\alias{FeatureSebyAuto}
\title{Title Automating the Feature Selection and benchmark judgement}
\usage{
FeatureSebyAuto(
  dataset = MLtestData,
  target = "judge",
  taskid = "MLtest",
  filtermethodlrn = "classif.rpart",
  nevals = 10,
  fselector = "random_search"
)
}
\arguments{
\item{dataset}{A dataframe of Machine Learing test dataset}

\item{target}{A character of column name in dataset}

\item{taskid}{A character of taskid}

\item{filtermethodlrn}{A filter selection by which lrn}

\item{nevals}{A number of nevals}

\item{fselector}{A character of fselector}
}
\value{

}
\description{
Title Automating the Feature Selection and benchmark judgement
}
