% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PepFeatureSelect.R
\name{FeatureSeByFilter}
\alias{FeatureSeByFilter}
\title{Title}
\usage{
FeatureSeByFilter(
  dataset = MLtestData,
  target = "judge",
  taskid = "MLtest",
  filtermethod = c("anova", "auc", "cmim", "disr", "find_correlation", "importance",
    "information_gain", "jmi", "jmim", "kruskal_test", "mim", "mrmr", "njmim",
    "performance", "permutation", "relief", "variance")
)
}
\arguments{
\item{dataset}{A dataframe of Machine Learing test dataset}

\item{target}{A character of column name in dataset}

\item{taskid}{A character of taskid}

\item{filtermethod}{A vector of filtering methods, each character should be one of mlr_filters$keys,which filter methods should be used,all filters method could be found by mlr_filters_,}
}
\value{

}
\description{
Title
}
