% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PepFeatureSelect.R
\name{FeatureSebyVariImpFil}
\alias{FeatureSebyVariImpFil}
\title{Title Feature Selection by Embedded Filter Methods}
\usage{
FeatureSebyVariImpFil(
  dataset = MLtestData,
  target = "judge",
  taskid = "FeatureSebyVariImpFil",
  filtermethodVI = c("classif.ranger", "classif.rpart", "classif.xgboost")
)
}
\arguments{
\item{dataset}{A dataframe of Machine Learing test dataset}

\item{target}{A character of column name in dataset}

\item{taskid}{A character of taskid}

\item{filtermethodVI}{A vectors of Learners With Embedded Filter Methods}
}
\value{

}
\description{
Title Feature Selection by Embedded Filter Methods
}
