% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PepFeatureSelect.R
\name{FeatureWrapperMethod}
\alias{FeatureWrapperMethod}
\title{Title Feature Selection by Wrapper Methods
Use this function in specific ML methods}
\usage{
FeatureWrapperMethod(
  dataset = MLtestData,
  target = "judge",
  taskid = "FeatureWrapperMethod",
  filtermethodlrn = "FeatureWrapperMethod",
  nevals = 20
)
}
\arguments{
\item{dataset}{A dataframe of Machine Learing test dataset}

\item{target}{A character of column name in dataset}

\item{taskid}{A character of taskid}

\item{filtermethodlrn}{A filter selection by which lrn}

\item{nevals}{A number of nevals}
}
\value{

}
\description{
Title Feature Selection by Wrapper Methods
Use this function in specific ML methods
}
