% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Neodataset}
\alias{Neodataset}
\title{formated neoantigens dataset for machine learning running}
\format{
A data frame with 6787 rows and 4 rows.
}
\usage{
Neodataset
}
\description{
A dataset of HLA-A*02:01 related simulated neopeptides.
}
\keyword{datasets}
