% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeoantigenML.R
\name{neoML.all}
\alias{neoML.all}
\title{Title glmnet/kknn/range/svm/xgboost training model}
\usage{
neoML.all(
  Neodataset = MLtestData,
  taskid = "neoML.glmnet",
  target = "judge",
  tunningkey = "classif.glmnet.rbv2",
  evalsnum = 20,
  innerparall = 4,
  outerparall = 3,
  innercv = 4,
  outercv = 3,
  measures = "classif.auc"
)
}
\arguments{
\item{Neodataset}{A data frame of feature selection dataset.}

\item{taskid}{A character of task id.}

\item{target}{A column name of prediction target in Neodataset}

\item{tunningkey}{A character of mlr_tuning_spaces c(classif.glmnet.rbv2,classif.kknn.rbv2,classif.ranger.rbv2,classif.svm.rbv2,classif.xgboost.default)}

\item{evalsnum}{the numbers of iterations evaluation}

\item{innerparall}{The number of inner parallelizations,default is 4}

\item{outerparall}{The number of outer parallelizations,default is 4}

\item{innercv}{The number of inner sampling in Nested sampling methods, default is 4}

\item{outercv}{The number of outer sampling in Nested sampling methods, default is 4}

\item{measures}{A character for Performance Measures detail could be found in https://mlr3book.mlr-org.com/appendix.html?q=measure#list-measures}
}
\value{
A resampling model with model.store
}
\description{
Title glmnet/kknn/range/svm/xgboost training model
}
\examples{
neoML.glmnet(MLtestData)
The evaluation of Tunning Space
}
