% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeoantigenML.R
\name{neoML}
\alias{neoML}
\title{Title}
\usage{
neoML(Neodataset)
}
\arguments{
\item{Neodataset}{A data frame of feature selection dataset}
}
\value{

}
\description{
Title
}
\examples{
neoML(Neodataset)
}
