% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeoantigenML.R
\name{neoML.rpart}
\alias{neoML.rpart}
\title{Title Classify by classif.rpart}
\usage{
neoML.rpart(Neodataset, taskid = "neoML", target = "judge", branchName = "nop")
}
\arguments{
\item{Neodataset}{A data frame of feature selection dataset.}

\item{taskid}{A character of task id.}

\item{target}{A column name of prediction target in Neodataset}

\item{branchName}{A character  Branch to use, which should be one of "nop", "pca", "scale"}
}
\value{
A matrix with accurate info
}
\description{
Title Classify by classif.rpart
}
\examples{
neoML.rpart(MLtestData)
}
