% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyofPepMutaion.R
\name{PepFeatureDesc.mutate}
\alias{PepFeatureDesc.mutate}
\title{Title peptide Feature Description}
\usage{
PepFeatureDesc.mutate(
  peptideSet,
  pair_peptideSet,
  featureSet = NULL,
  coreN = parallel::detectCores(logical = F),
  fragLenSet = 2:8,
  CutStart = NULL,
  Cutend = NULL,
  peptideSet.pos
)
}
\description{
Title peptide Feature Description
}
