% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepKmerFeature.R
\name{PepFeatureDesc}
\alias{PepFeatureDesc}
\title{Title peptide Feature Description}
\usage{
PepFeatureDesc(
  peptideSet,
  fragLenSet = 3:8,
  featureSet = NULL,
  coreN = parallel::detectCores(logical = F),
  CutStart = NULL,
  Cutend = NULL
)
}
\description{
Title peptide Feature Description
}
