% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepKmerFeature.R
\name{PepFrag}
\alias{PepFrag}
\title{Create fragment of peptide set}
\usage{
PepFrag(
  peptideSet = peptides,
  coreN = parallel::detectCores(logical = F),
  fragLenSet = 3:8,
  CutStart = NULL,
  Cutend = NULL
)
}
\value{
A vector with all fragment of peptide set
}
\description{
Create fragment of peptide set
}
\examples{
PepFragResult <- PepFrag(peptideSet = peptides.allmouse,fragLenSet = 3:8)
}
