% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepKmerFeature.R
\name{PepFragSign}
\alias{PepFragSign}
\title{Title PepFraSign}
\usage{
PepFragSign(
  PepFrag = PepFragResult,
  peptides = peptides.allmouse,
  JudgeMatrix = pepJudgeMatrix,
  prop = 0.01,
  coreN = parallel::detectCores(logical = F),
  Fake = FALSE,
  FakeMatrix = NULL
)
}
\value{
d datatable with significant enrichment fragment
}
\description{
Title PepFraSign
}
\examples{
pepJudgeMatrix <- deredundancy.allmouse \%>\% column_to_rownames(var = "description")
pepFragSignResult <- PepFragSign()
}
