% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepVisualization.R
\name{pepFragVis}
\alias{pepFragVis}
\title{Title visualization for fragment of peptides}
\usage{
pepFragVis(
  PepFragChoose,
  peptideSet = peptides.allhuman,
  pepJudgeMatrix = PepjudgeMatrixexam,
  outPath = "./result/Kmer_result/",
  prefix = "Kmer100_"
)
}
\arguments{
\item{PepFragChoose}{a vector with all signaficant enriched fragment of peptides}

\item{peptideSet}{A vector of all peptides with uesed to generate fragment}

\item{pepJudgeMatrix}{a dataframe with judge info of peptides}

\item{outPath}{output path}

\item{prefix}{a character of prefix of output name}
}
\value{

}
\description{
Title visualization for fragment of peptides
}
