% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pepStat.R
\name{pepLength}
\alias{pepLength}
\title{Length Statistics}
\usage{
pepLength(pepSeq)
}
\arguments{
\item{pepSeq}{a vector of sequence,which could get from by}
}
\value{
ggbar A bar plot from ggpplot
}
\description{
Length Statistics
}
\examples{
pepSeq <- Neodataset$wild_Peptide
pepLength(pepSeq = pepSeq)

}
