% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptides}
\alias{peptides}
\title{positive and negative neopeptides from  netMHCpan}
\format{
A data frame with 755 rows and 7 variables in positive; A negative data frame with 6032 rows and  columns
}
\usage{
peptides
}
\description{
A example peptides vector
}
\keyword{datasets}
