% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PepDataVis.R
\name{pepVisLength}
\alias{pepVisLength}
\title{Visualization of length statistics in peptides.}
\usage{
pepVisLength(pepSeq)
}
\arguments{
\item{pepSeq}{a vector of sequence.}
}
\value{
ggbar A bar plot from ggpplot.
}
\description{
Visualization of length statistics in peptides.
}
\examples{
pepSeq <- Neodataset$wild_Peptide
pepLength(pepSeq = pepSeq)

}
