% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEva.R
\name{PerformanceEva}
\alias{PerformanceEva}
\title{Title Performance evaluation of single machine learning model.}
\usage{
PerformanceEva(
  task,
  learner,
  resamplemethod = rsmp("cv", folds = 3),
  bmr_small
)
}
\arguments{
\item{task}{A task object created by mlr3}

\item{resamplemethod}{A resample object created by mlr3}

\item{bmr_small}{A single benchmark object such as  bmr_small = bmr$clone()$filter(task_id = tasks\link{1})}

\item{leaner}{A learner object created by mlr3}
}
\value{
A list with all performance results.
}
\description{
Title Performance evaluation of single machine learning model.
}
\examples{
data("Sonar", package = "mlbench")
task = as_task_classif(Sonar, target = "Class", positive = "M")
learner = lrn("classif.rpart", predict_type = "prob")
PerformanceEva(task = task,learner = learner)
}
