% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureCorrelation.R
\name{Pointbiserial}
\alias{Pointbiserial}
\title{Title}
\usage{
Pointbiserial(FeaMatrix, judge)
}
\arguments{
\item{FeaMatrix}{a dataframe of Featrue matrix}

\item{judge}{a vector of judge}
}
\value{

}
\description{
Title
}
