% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropertyofPepSingle.R
\name{PropertyofPepSingle}
\alias{PropertyofPepSingle}
\title{Title Properties of whole peptides}
\usage{
PropertyofPepSingle(
  peptides,
  autoCor = TRUE,
  lag = 1,
  property = AAdata$Hydrophobicity$KyteDoolittle,
  property1 = AAdata$Hydrophobicity$KyteDoolittle,
  property2 = AAdata$Hydrophobicity$Eisenberg,
  center = TRUE,
  pH = 7,
  pKscale = "Lehninger",
  scale = "KyteDoolittle",
  monoisotopic = FALSE,
  avgScale = "expasy",
  aaShift = NULL,
  charge = 2,
  pKscale.PI = "EMBOSS"
)
}
\arguments{
\item{peptides}{A vector of peptides with same length}

\item{autoCor}{A logic value TRUE or FALSE, meaning to calculate autoCovariance value or not.}

\item{lag}{A value for a lag, the max value is equal to the length of shortest peptide minus one. detail ref 'Peptides' package.}

\item{property}{A property to use as value to be correlated.}

\item{property1}{A property to use as value to evaluate the cross-covariance.}

\item{property2}{A property to use as value to evaluate the cross-covariance.}

\item{center}{A logic value TRUE or FALSE if the property must be centered}

\item{pH}{A pH value in charge function.}

\item{pKscale}{A character string specifying the pKa scale to be used; must be one of "Bjellqvist","Dawson", "EMBOSS", "Lehninger", "Murray", "Rodwell", "Sillero", "Solomon"or "Stryer"}

\item{scale}{A character string specifying the hydophobicity scale to be used; must be one of "Aboderin", "AbrahamLeo", "Argos", "BlackMould", "BullBreese", "Casari","Chothia", "Cid", "Cowan3.4", "Cowan7.5", "Eisenberg", "Engelman", "Fasman","Fauchere", "Goldsack", "Guy", "HoppWoods", "Janin", "Jones", "Juretic","Kidera", "Kuhn", "KyteDoolittle", "Levitt", "Manavalan", "Miyazawa","Parker", "Ponnuswamy", "Prabhakaran", "Rao", "Rose", "Roseman", "Sweet","Tanford", "Welling", "Wilson", "Wolfenden", "Zimmerman", "interfaceScale_pH8","interfaceScale_pH2", "octanolScale_pH8", "octanolScale_pH2", "oiScale_pH8"or "oiScale_pH2".}

\item{monoisotopic}{A logical value 'TRUE' or 'FALSE' indicating if monoisotopic weights of aminoacids should be used}

\item{avgScale}{Set the mass scale to use for average weight only (if ’monoisotopic == FALSE’).Accepts "expasy" (default) or "mascot".}

\item{aaShift}{Define the mass difference in Dalton of given amino acids as a named vector.Use the amino acid one letter code as names and the mass shift in Dalton as values.}

\item{charge}{The net charge for which the m/z should be calculated}

\item{pKscale.PI}{A character string specifying the pK scale in PI function to be used; must be one of "Bjellqvist","EMBOSS", "Murray", "Sillero", "Solomon", "Stryer", "Lehninger", "Dawson"or "Rodwell"}
}
\value{
a dataframe with all properties.
}
\description{
Title Properties of whole peptides
}
\examples{
Neodataset.length <- Neodataset \%>\% 
mutate(Length = str_length(wild_Peptide)) \%>\% 
filter(Length == 9) \%>\% 
unique()
peptides <- unique(Neodataset.length$wild_Peptide) 
PropertyofPep(peptides)
}
