% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEva.R
\name{BenchmarkEva}
\alias{BenchmarkEva}
\title{Title Benchmark calculation
In particular unique combination of Task,Learner and Resampling.}
\usage{
BenchmarkEva(tasks, learners, resamplings = "none")
}
\arguments{
\item{tasks}{A vector of task}

\item{learners}{A list of learners}

\item{resamplings}{A resamplings methods}
}
\value{

}
\description{
Title Benchmark calculation
In particular unique combination of Task,Learner and Resampling.
}
\examples{
design = benchmark_grid(
tasks = tsks(c("spam", "german_credit", "sonar")),
learners = lrns(c("classif.glmnet", "classif.rpart", "classif.featureless"),
               predict_type = "prob", predict_sets = c("train", "test")),
resamplings = rsmps("cv", folds = 3)
classif.glmnet.rbv2,classif.kknn.rbv2,classif.ranger.rbv2,classif.svm.rbv2,classif.xgboost.rbv2,
}
