% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformanceEva.R
\name{BenchmarkEva}
\alias{BenchmarkEva}
\title{Title Benchmark calculation
In particular unique combination of Task,Learner and Resampling.}
\usage{
BenchmarkEva(tasks, learners, reamplings)
}
\arguments{
\item{tasks}{A vector of task}

\item{learners}{A vector of learners}

\item{reamplings}{A resamplings object. detail could be found in example}
}
\value{

}
\description{
Title Benchmark calculation
In particular unique combination of Task,Learner and Resampling.
}
\examples{
design = benchmark_grid(
tasks = tsks(c("spam", "german_credit", "sonar")),
learners = lrns(c("classif.ranger", "classif.rpart", "classif.featureless"),
               predict_type = "prob", predict_sets = c("train", "test")),
resamplings = rsmps("cv", folds = 3)
}
